pyproject.toml
[tool.poetry]
name = "ninja-api"
version = "0.1.0"
description = "Django-Ninja API (using Django)"
authors = ["Eduardo Rezaghi <eduardo.rezaghi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-ninja = "^1.3.0"
pytest-django = "^4.9.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.32.0"
mypy = "^1.15.0"
ruff = "^0.9.5"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -v -p no:warnings"
DJANGO_SETTINGS_MODULE = "bookstore.settings"
testpaths = [
    "tests",
    "integration",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

